requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cu117
transformers==4.27.2
coclust==0.2.1
scikit-learn==1.0.2
umap-learn==0.5.2
pandas